requirements.txt
python-dotenv
openai
langchain
elevenlabs
fastapi
uvicorn
twilio
streamlit
